Rename libevent_watcher.{h,cc} to event_watcher.{h,cc}
