+++ apps/evmc/mctypes.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "evpp/libevent_watcher.h"
+#include "evpp/event_watcher.h"
 #include "evpp/event_loop_thread.h"
 #include "evpp/buffer.h"
 #include "evpp/tcp_conn.h"
