+++ apps/evmc/memcache_client.h
@@ -7,7 +7,7 @@
 #include "evpp/tcp_conn.h"
 #include "evpp/tcp_client.h"
 
-#include "evpp/libevent_watcher.h"
+#include "evpp/event_watcher.h"
 #include "evpp/event_loop.h"
 #include "evpp/event_loop_thread_pool.h"
