+++ benchmark/ioevent/evpp/evpp_ioevent_pipe_watcher.cc
@@ -5,7 +5,7 @@
 
 #include <evpp/event_loop.h>
 #include <evpp/fd_channel.h>
-#include <evpp/libevent_watcher.h>
+#include <evpp/event_watcher.h>
 
 #include <getopt.h>
