+++ benchmark/ioevent/fd_channel_vs_pipe_event_watcher/fd_channel_vs_pipe_event_watcher.cc
@@ -3,7 +3,7 @@
 #include <evpp/event_loop.h>
 #include <evpp/fd_channel.h>
 #include <evpp/libevent_headers.h>
-#include <evpp/libevent_watcher.h>
+#include <evpp/event_watcher.h>
 
 #include <getopt.h>
