+++ evpp/dns_resolver.cc
@@ -1,7 +1,7 @@
 
 #include "evpp/dns_resolver.h"
 #include "evpp/event_loop.h"
-#include "evpp/libevent_watcher.h"
+#include "evpp/event_watcher.h"
 #include "evpp/libevent_headers.h"
 
 namespace evpp {
