+++ evpp/event_watcher.cc
@@ -2,8 +2,8 @@
 
 #include <string.h>
 
-#include "evpp/libevent_watcher.h"
 #include "evpp/libevent_headers.h"
+#include "evpp/event_watcher.h"
 #include "evpp/event_loop.h"
 
 namespace evpp {
