+++ evpp/http/http_server.cc
@@ -3,7 +3,7 @@
 #include <future>
 
 #include "evpp/libevent_headers.h"
-#include "evpp/libevent_watcher.h"
+#include "evpp/event_watcher.h"
 #include "evpp/event_loop.h"
 #include "evpp/event_loop_thread.h"
 #include "evpp/event_loop_thread_pool.h"
