+++ evpp/http/service.cc
@@ -1,7 +1,7 @@
 #include "service.h"
 
 #include "evpp/libevent_headers.h"
-#include "evpp/libevent_watcher.h"
+#include "evpp/event_watcher.h"
 #include "evpp/event_loop.h"
 
 namespace evpp {
