+++ examples/signal/signal02.cc
@@ -1,5 +1,5 @@
 #include <signal.h>
-#include <evpp/libevent_watcher.h>
+#include <evpp/event_watcher.h>
 #include <evpp/event_loop.h>
 
 #include "examples/winmain-inl.h"
