+++ test/event_loop_test.cc
@@ -2,7 +2,7 @@
 #include "test_common.h"
 
 #include <evpp/libevent_headers.h>
-#include <evpp/libevent_watcher.h>
+#include <evpp/event_watcher.h>
 #include <evpp/event_loop.h>
 #include <evpp/timestamp.h>
