+++ test/tcp_client_reconnect_test.cc
@@ -1,7 +1,7 @@
 #include "test_common.h"
 
 #include <evpp/libevent_headers.h>
-#include <evpp/libevent_watcher.h>
+#include <evpp/event_watcher.h>
 #include <evpp/event_loop.h>
 #include <evpp/event_loop_thread.h>
 #include <evpp/tcp_server.h>
