+++ vsprojects/libevpp.vcxproj
@@ -30,7 +30,7 @@
     <ClCompile Include="..\evpp\inner_pre.cc" />
     <ClCompile Include="..\evpp\invoke_timer.cc" />
     <ClCompile Include="..\evpp\libevent_compact.cc" />
-    <ClCompile Include="..\evpp\libevent_watcher.cc" />
+    <ClCompile Include="..\evpp\event_watcher.cc" />
     <ClCompile Include="..\evpp\listener.cc" />
     <ClCompile Include="..\evpp\sockets.cc" />
     <ClCompile Include="..\evpp\tcp_client.cc" />
@@ -63,7 +63,7 @@
     <ClInclude Include="..\evpp\inner_pre.h" />
     <ClInclude Include="..\evpp\invoke_timer.h" />
     <ClInclude Include="..\evpp\libevent_headers.h" />
-    <ClInclude Include="..\evpp\libevent_watcher.h" />
+    <ClInclude Include="..\evpp\event_watcher.h" />
     <ClInclude Include="..\evpp\listener.h" />
     <ClInclude Include="..\evpp\log_config.h" />
     <ClInclude Include="..\evpp\memmem.h" />
