+++ vsprojects/libevpp.vcxproj.filters
@@ -64,9 +64,6 @@
     <ClCompile Include="..\evpp\buffer.cc">
       <Filter>common</Filter>
     </ClCompile>
-    <ClCompile Include="..\evpp\libevent_watcher.cc">
-      <Filter>evloop</Filter>
-    </ClCompile>
     <ClCompile Include="..\evpp\tcp_conn.cc">
       <Filter>tcp</Filter>
     </ClCompile>
@@ -114,6 +111,9 @@
     </ClCompile>
     <ClCompile Include="..\evpp\udp\sync_udp_client.cc">
       <Filter>udp</Filter>
+    </ClCompile>
+    <ClCompile Include="..\evpp\event_watcher.cc">
+      <Filter>evloop</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -183,9 +183,6 @@
     <ClInclude Include="..\evpp\utility.h">
       <Filter>common</Filter>
     </ClInclude>
-    <ClInclude Include="..\evpp\libevent_watcher.h">
-      <Filter>evloop</Filter>
-    </ClInclude>
     <ClInclude Include="..\evpp\tcp_callbacks.h">
       <Filter>tcp</Filter>
     </ClInclude>
@@ -248,6 +245,9 @@
     </ClInclude>
     <ClInclude Include="..\evpp\server_status.h">
       <Filter>common</Filter>
+    </ClInclude>
+    <ClInclude Include="..\evpp\event_watcher.h">
+      <Filter>evloop</Filter>
     </ClInclude>
   </ItemGroup>
 </Project>
